backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.2
xgboost==2.0.3
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.0
httpx==0.27.0
python-dotenv==1.0.1
scipy==1.13.0
